move 'load_aea_package' from 'aea.helpers' to 'aea.component.base'
This caused circular dependency, since the aea.helpers module imported
from aea.configurations.
