+++ aea/helpers/base.py
@@ -40,8 +40,6 @@ from dotenv import load_dotenv
 
 import yaml
 
-from aea.configurations.base import ComponentConfiguration
-
 logger = logging.getLogger(__name__)
 
 
@@ -167,44 +165,6 @@ def locate(path: str) -> Any:
         except AttributeError:
             return None
     return object_
-
-
-def load_aea_package(configuration: ComponentConfiguration) -> None:
-    """
-    Load the AEA package.
-
-    It adds all the __init__.py modules into `sys.modules`.
-
-    :param configuration: the configuration object.
-    :return: None
-    """
-    dir_ = configuration.directory
-    assert dir_ is not None
-
-    # patch sys.modules with dummy modules
-    prefix_root = "packages"
-    prefix_author = prefix_root + f".{configuration.author}"
-    prefix_pkg_type = prefix_author + f".{configuration.component_type.to_plural()}"
-    prefix_pkg = prefix_pkg_type + f".{configuration.name}"
-    sys.modules[prefix_root] = types.ModuleType(prefix_root)
-    sys.modules[prefix_author] = types.ModuleType(prefix_author)
-    sys.modules[prefix_pkg_type] = types.ModuleType(prefix_pkg_type)
-
-    for subpackage_init_file in dir_.rglob("__init__.py"):
-        parent_dir = subpackage_init_file.parent
-        relative_parent_dir = parent_dir.relative_to(dir_)
-        if relative_parent_dir == Path("."):
-            # this handles the case when 'subpackage_init_file'
-            # is path/to/package/__init__.py
-            import_path = prefix_pkg
-        else:
-            import_path = prefix_pkg + "." + ".".join(relative_parent_dir.parts)
-
-        spec = importlib.util.spec_from_file_location(import_path, subpackage_init_file)
-        module = importlib.util.module_from_spec(spec)
-        sys.modules[import_path] = module
-        logger.debug(f"loading {import_path}: {module}")
-        spec.loader.exec_module(module)  # type: ignore
 
 
 def load_module(dotted_path: str, filepath: Path) -> types.ModuleType:
