+++ aea/protocols/base.py
@@ -31,14 +31,13 @@ from typing import Any, Dict, Optional, Tuple, Type, cast
 
 from google.protobuf.struct_pb2 import Struct
 
-from aea.components.base import Component
+from aea.components.base import Component, load_aea_package
 from aea.configurations.base import (
     ComponentConfiguration,
     ComponentType,
     ProtocolConfig,
     PublicId,
 )
-from aea.helpers.base import load_aea_package
 
 logger = logging.getLogger(__name__)
